fix useFlexgetStream
